added confirmation when deleting and editing patient

+++ HealthEdgeFrontend/healthedgefrontend/src/pages/Patients.js
@@ -2,7 +2,8 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
 import PatientCard from '../components/patients/PatientCard';
-
+import { confirmAlert } from 'react-confirm-alert';
+import 'react-confirm-alert/src/react-confirm-alert.css';
 const Patients = () => {
   const [patients, setPatients] = useState([]);
   const navigate = useNavigate();
@@ -24,13 +25,28 @@ const Patients = () => {
     navigate(`/edit-patient/${patientId}`);
   };
 
-  const handleDelete = async (patientId) => {
-    try {
-      await axios.delete(`http://localhost:5239/patients/delete/${patientId}`);
-      setPatients(patients.filter((patient) => patient.id !== patientId));
-    } catch (error) {
-      console.error('Failed to delete patient:', error);
-    }
+  const handleDelete = (patientId) => {
+    confirmAlert({
+      title: 'Confirm to delete',
+      message: 'Are you sure you want to delete this patient?',
+      buttons: [
+        {
+          label: 'Yes',
+          onClick: async () => {
+            try {
+              await axios.delete(`http://localhost:5239/patients/delete/${patientId}`);
+              setPatients(patients.filter((patient) => patient.id !== patientId));
+            } catch (error) {
+              console.error('Failed to delete patient:', error);
+            }
+          }
+        },
+        {
+          label: 'No',
+          onClick: () => {}
+        }
+      ]
+    });
   };
 
   return (
